Added a health bar for Enemy.java
Added images that indicate health of snowman (# of snow balls = health). Furthermore, made a method called damage() that will remove reduce the health of snowman by 1, remove the snowball from world and change the image of snowman to show health. If the health reaches 0, the snowman is removed.

Also, added a method in MyWorld called SpawnSnowman() that creates a new Snowman object that is placed randomly on the world. This is called whenever a snowman dies.